Move create_source_list et al. to a new module
